enhance: avoid depending on cgo for pkg module
