+++ internal/util/initcore/hooks.go
@@ -1,0 +1,32 @@
+// Licensed to the LF AI & Data foundation under one
+// or more contributor license agreements. See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership. The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License. You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package initcore
+
+import "github.com/milvus-io/milvus/pkg/v2/util/paramtable"
+
+func init() {
+	paramtable.UpdateLogLevel = UpdateLogLevel
+	paramtable.UpdateIndexSliceSize = UpdateIndexSliceSize
+	paramtable.UpdateHighPriorityThreadCoreCoefficient = UpdateHighPriorityThreadCoreCoefficient
+	paramtable.UpdateMiddlePriorityThreadCoreCoefficient = UpdateMiddlePriorityThreadCoreCoefficient
+	paramtable.UpdateLowPriorityThreadCoreCoefficient = UpdateLowPriorityThreadCoreCoefficient
+	paramtable.UpdateDefaultExprEvalBatchSize = UpdateDefaultExprEvalBatchSize
+	paramtable.UpdateDefaultOptimizeExprEnable = UpdateDefaultOptimizeExprEnable
+	paramtable.UpdateDefaultJSONKeyStatsCommitInterval = UpdateDefaultJSONKeyStatsCommitInterval
+	paramtable.UpdateDefaultGrowingJSONKeyStatsEnable = UpdateDefaultGrowingJSONKeyStatsEnable
+	paramtable.UpdateDefaultConfigParamTypeCheck = UpdateDefaultConfigParamTypeCheck
+}
